Reworking
Reworking an editing some of the lines to make it better. Making it
instead have a list that will hold current parent branch. Will check for
cases and then create a new node depending on the case. Will set the
right and left of the parent. Will then go through and set the parent
equal to the current node.  Will then add  the left and right child to a
list. The child list will then replace the parent list.